+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 - Fix up handling of tags without an end of line anchor (#defines?)
+- Improved console logging
 
 ## [0.0.2]